FEATURE: Add dataSrc method to image source interface
This allows to render images with a data src to avoid a separate request.
The dataSrc is generated synchronously which may cause performance issues.

To achieve this the generation of dummy images was extracted into the DummyImageGenerator
for beeing able to crate synchonous dummy images.

Also the test matrix was adjusted to include recent and upcoming php and Neos versions
and removes support of outdated Neos versions.
