+++ composer.json
@@ -4,8 +4,8 @@
     "name": "sitegeist/kaleidoscope",
     "license": "GPL-3.0-or-later",
     "require": {
-        "neos/neos": "^5.0 || ^7.0 || ^8.0 || ^9.0 || dev-master",
-        "neos/fusion-afx": "^1.2 || ^7.0 || ^8.0 || ^9.0 || dev-master",
+        "neos/neos": "^7.0 || ^8.0 || ^9.0 || dev-master",
+        "neos/fusion-afx": "^7.0 || ^8.0 || ^9.0 || dev-master",
         "neos/media": "*",
         "neos/imagine": "*"
     },
